Changed style swapping to work on styles that come from the local document OR a library. - Expanded which styles are included in making matches from just local styles to any style. - Renamed the menu items and dialog explanations to match the new behavior, and to make more clear. - Renamed the menu item for swapping color variables to match the naming used for styles. - Updated readme to match the above changes.
The purpose of this change is to make it possible to use Automate to change where you keep your
styles. When splitting a single library into two, you need to update both the library and files that
use that library. For the library being split, Automate works fine as-is. But for the files that use
the library, we need a way to change from using the original library to using the new one. This
change makes it possible to use the existing feature in Automate for both parts of the task.

TODO:
- Chinese translations in manifest_zh.json and readme_zh.md need to be done. (I’m not qualified!)